Changes name of Project to ProjectArgs.
This was done for clarity. Previously there was a naming collision
between the struct in the API crate and the struct
written for command line argument parsing. I renamed the latter
to `ProjectArgs` (and `Projects` to `ProjectArgs`). This freed up
the `Project` name for the structs from the oxidoist-api crate, which
hold project data from Todoist's API.